Remove RCPSPModelSpecialConstraints
- We merge it with RPCPModel
- We use do_special_constraints attribute and includes_special_constraint
method to knwow when to apply special constraints.
- We merge RCPSPSolutionSpecial with RCPSPSolution and use
  self.problem.do_special_constraints to behave differently
- We put stuff that can be reused in a separate module core.py
- We put functions used in case of special constraints in
  rcpsp_model.py to avoid circular imports
