+++ discrete_optimization/rcpsp/solver/rcpsp_lp_solver.py
@@ -24,8 +24,8 @@ from discrete_optimization.generic_tools.lp_tools import (
 from discrete_optimization.generic_tools.result_storage.result_storage import (
     ResultStorage,
 )
+from discrete_optimization.rcpsp.core import PartialSolution
 from discrete_optimization.rcpsp.rcpsp_model import (
-    PartialSolution,
     RCPSPModel,
     RCPSPSolution,
     Solution,
