+++ discrete_optimization/rcpsp/specialized_rcpsp/rcpsp_specialized_constraints.py
@@ -6,9 +6,8 @@ import logging
 import random
 from copy import deepcopy
 from functools import partial
-from typing import Dict, Hashable, List, Optional, Set, Tuple, Type, Union
+from typing import Dict, Hashable, List, Optional, Set, Type, Union
 
-import networkx as nx
 import numpy as np
 
 from discrete_optimization.generic_tools.do_problem import (
@@ -20,6 +19,7 @@ from discrete_optimization.generic_tools.do_problem import (
     Solution,
     TypeObjective,
 )
+from discrete_optimization.rcpsp.core import SpecialConstraintsDescription
 from discrete_optimization.rcpsp.fast_function_rcpsp import (
     compute_mean_ressource,
     sgs_fast_partial_schedule_preemptive,
@@ -27,11 +27,7 @@ from discrete_optimization.rcpsp.fast_function_rcpsp import (
     sgs_fast_preemptive_minduration,
     sgs_fast_preemptive_some_special_constraints,
 )
-from discrete_optimization.rcpsp.rcpsp_model import (
-    PartialSolution,
-    RCPSPModel,
-    RCPSPSolution,
-)
+from discrete_optimization.rcpsp.rcpsp_model import RCPSPModel, RCPSPSolution
 from discrete_optimization.rcpsp.rcpsp_model_preemptive import (
     RCPSPModelPreemptive,
     RCPSPSolutionPreemptive,
@@ -39,288 +35,6 @@ from discrete_optimization.rcpsp.rcpsp_model_preemptive import (
 from discrete_optimization.rcpsp.rcpsp_utils import intersect
 
 logger = logging.getLogger(__name__)
-
-
-class SpecialConstraintsDescription(PartialSolution):
-    def __init__(
-        self,
-        task_mode: Dict[Hashable, int] = None,
-        start_times: Dict[Hashable, int] = None,
-        end_times: Dict[Hashable, int] = None,
-        start_times_window: Dict[
-            Hashable, Tuple[Union[int, None], Union[int, None]]
-        ] = None,
-        end_times_window: Dict[
-            Hashable, Tuple[Union[int, None], Union[int, None]]
-        ] = None,
-        partial_permutation: List[Hashable] = None,
-        list_partial_order: List[List[Hashable]] = None,
-        start_together: List[Tuple[Hashable, Hashable]] = None,
-        start_at_end: List[Tuple[Hashable, Hashable]] = None,
-        start_at_end_plus_offset: List[Tuple[Hashable, Hashable, int]] = None,
-        start_after_nunit: List[Tuple[Hashable, Hashable, int]] = None,
-        disjunctive_tasks: List[Tuple[Hashable, Hashable]] = None,
-    ):
-        super().__init__(
-            task_mode=task_mode,
-            start_times=start_times,
-            end_times=end_times,
-            partial_permutation=partial_permutation,
-            list_partial_order=list_partial_order,
-            start_together=start_together,
-            start_at_end=start_at_end,
-            start_at_end_plus_offset=start_at_end_plus_offset,
-            start_after_nunit=start_after_nunit,
-            start_times_window=start_times_window,
-            end_times_window=end_times_window,
-            disjunctive_tasks=disjunctive_tasks,
-        )
-        if self.start_times_window is None:
-            self.start_times_window = {}
-        if self.end_times_window is None:
-            self.end_times_window = {}
-        if self.start_together is None:
-            self.start_together = []
-        if self.start_at_end is None:
-            self.start_at_end = []
-        if self.start_at_end_plus_offset is None:
-            self.start_at_end_plus_offset = []
-        if self.start_after_nunit is None:
-            self.start_after_nunit = []
-        if self.disjunctive_tasks is None:
-            self.disjunctive_tasks = []
-        self.dict_start_together = {}
-        self.graph_start_together = nx.Graph()
-        for i, j in self.start_together:
-            if i not in self.graph_start_together:
-                self.graph_start_together.add_node(i)
-            if j not in self.graph_start_together:
-                self.graph_start_together.add_node(j)
-            self.graph_start_together.add_edge(i, j)
-        self.components = [
-            c for c in nx.connected_components(self.graph_start_together)
-        ]
-        for c in self.components:
-            for j in c:
-                self.dict_start_together[j] = set([k for k in c if k != j])
-        self.dict_start_at_end = {}
-        self.dict_start_at_end_reverse = {}
-        for i, j in self.start_at_end:
-            if i not in self.dict_start_at_end:
-                self.dict_start_at_end[i] = set()
-            if j not in self.dict_start_at_end_reverse:
-                self.dict_start_at_end_reverse[j] = set()
-            self.dict_start_at_end[i].add(j)
-            self.dict_start_at_end_reverse[j].add(i)
-
-        self.dict_start_at_end_offset = {}
-        self.dict_start_at_end_offset_reverse = {}
-        for i, j, off in self.start_at_end_plus_offset:
-            if i not in self.dict_start_at_end_offset:
-                self.dict_start_at_end_offset[i] = {}
-            if j not in self.dict_start_at_end_offset_reverse:
-                self.dict_start_at_end_offset_reverse[j] = {}
-            self.dict_start_at_end_offset_reverse[j][i] = off
-            self.dict_start_at_end_offset[i][j] = off
-
-        self.dict_start_after_nunit = {}
-        self.dict_start_after_nunit_reverse = {}
-        for i, j, off in self.start_after_nunit:
-            if i not in self.dict_start_after_nunit:
-                self.dict_start_after_nunit[i] = {}
-            if j not in self.dict_start_after_nunit_reverse:
-                self.dict_start_after_nunit_reverse[j] = {}
-            self.dict_start_after_nunit[i][j] = off
-            self.dict_start_after_nunit_reverse[j][i] = off
-
-        self.dict_disjunctive = {}
-        for i, j in self.disjunctive_tasks:
-            if i not in self.dict_disjunctive:
-                self.dict_disjunctive[i] = set()
-            if j not in self.dict_disjunctive:
-                self.dict_disjunctive[j] = set()
-            self.dict_disjunctive[i].add(j)
-            self.dict_disjunctive[j].add(i)
-
-
-class RCPSPSolutionSpecial(RCPSPSolution):
-    def copy(self) -> "RCPSPSolutionSpecial":
-        return RCPSPSolutionSpecial(
-            problem=self.problem,
-            rcpsp_permutation=deepcopy(self.rcpsp_permutation),
-            rcpsp_modes=deepcopy(self.rcpsp_modes),
-            rcpsp_schedule=deepcopy(self.rcpsp_schedule),
-            rcpsp_schedule_feasible=self.rcpsp_schedule_feasible,
-            standardised_permutation=self.standardised_permutation,
-            fast=self.fast,
-        )
-
-    def lazy_copy(self) -> "RCPSPSolutionSpecial":
-        return RCPSPSolutionSpecial(
-            problem=self.problem,
-            rcpsp_permutation=self.rcpsp_permutation,
-            rcpsp_modes=self.rcpsp_modes,
-            rcpsp_schedule=self.rcpsp_schedule,
-            rcpsp_schedule_feasible=self.rcpsp_schedule_feasible,
-            standardised_permutation=self.standardised_permutation,
-            fast=self.fast,
-        )
-
-    def generate_permutation_from_schedule(self):
-        sorted_task = [
-            self.problem.index_task[i] - 1
-            for i in sorted(
-                self.rcpsp_schedule, key=lambda x: self.rcpsp_schedule[x]["start_time"]
-            )
-        ]
-        sorted_task.remove(-1)
-        sorted_task.remove(max(sorted_task))
-        return sorted_task
-
-    def generate_schedule_from_permutation_serial_sgs_2(
-        self,
-        current_t=0,
-        completed_tasks=None,
-        scheduled_tasks_start_times=None,
-        do_fast=True,
-    ):
-        if completed_tasks is None:
-            completed_tasks = {}
-        if scheduled_tasks_start_times is None:
-            scheduled_tasks_start_times = None
-        (
-            schedule,
-            feasible,
-        ) = generate_schedule_from_permutation_serial_sgs_partial_schedule(
-            solution=self,
-            current_t=current_t,
-            completed_tasks=completed_tasks,
-            scheduled_tasks_start_times=scheduled_tasks_start_times,
-            rcpsp_problem=self.problem,
-        )
-        self.rcpsp_schedule = schedule
-        self._schedule_to_recompute = False
-
-
-class RCPSPModelSpecialConstraints(RCPSPModel):
-    def __init__(
-        self,
-        resources: Union[Dict[str, int], Dict[str, List[int]]],
-        non_renewable_resources: List[str],
-        mode_details: Dict[Hashable, Dict[Union[str, int], Dict[str, int]]],
-        successors: Dict[Union[int, str], List[Union[str, int]]],
-        horizon,
-        special_constraints: SpecialConstraintsDescription = None,
-        relax_the_start_at_end: bool = True,
-        tasks_list: List[Union[int, str]] = None,
-        source_task=None,
-        sink_task=None,
-        name_task: Dict[int, str] = None,
-        **args
-    ):
-        self.special_constraints = special_constraints
-        self.do_special_constraints = self.special_constraints is not None
-        if self.special_constraints is None:
-            self.special_constraints = SpecialConstraintsDescription()
-        super().__init__(
-            resources,
-            non_renewable_resources,
-            mode_details,
-            successors,
-            horizon,
-            tasks_list=tasks_list,
-            source_task=source_task,
-            sink_task=sink_task,
-            name_task=name_task,
-        )
-        self.predecessors_dict = {task: [] for task in self.successors}
-        for task in self.successors:
-            for stask in self.successors[task]:
-                self.predecessors_dict[stask] += [task]
-        if self.do_special_constraints:
-            for t1, t2 in self.special_constraints.start_at_end:
-                if t2 not in self.successors[t1]:
-                    self.successors[t1].append(t2)
-            for t1, t2, off in self.special_constraints.start_at_end_plus_offset:
-                if t2 not in self.successors[t1]:
-                    self.successors[t1].append(t2)
-            for t1, t2 in self.special_constraints.start_together:
-                for predt1 in self.predecessors_dict[t1]:
-                    if t2 not in self.successors[predt1]:
-                        self.successors[predt1] += [t2]
-                for predt2 in self.predecessors_dict[t2]:
-                    if t1 not in self.successors[predt2]:
-                        self.successors[predt2] += [t1]
-        self.graph = self.compute_graph(compute_predecessors=True)
-        self.predecessors = self.graph.predecessors_dict
-        self.sgs_func = generate_schedule_from_permutation_serial_sgs
-        self.sgs_func_partial = (
-            generate_schedule_from_permutation_serial_sgs_partial_schedule
-        )
-        self.relax_the_start_at_end = relax_the_start_at_end
-
-    def has_special_constraints(self):
-        return self.do_special_constraints
-
-    def is_preemptive(self) -> bool:
-        return False
-
-    def evaluate_from_encoding(self, int_vector, encoding_name):
-        if encoding_name == "rcpsp_permutation":
-            single_mode_list = [1 for i in range(self.n_jobs_non_dummy)]
-            rcpsp_sol = RCPSPSolutionSpecial(
-                problem=self, rcpsp_permutation=int_vector, rcpsp_modes=single_mode_list
-            )
-            objectives = self.evaluate(rcpsp_sol)
-            return objectives
-        return None
-
-    def evaluate_function(self, rcpsp_sol: RCPSPSolution):
-        if rcpsp_sol._schedule_to_recompute:
-            rcpsp_sol.generate_schedule_from_permutation_serial_sgs()
-        makespan = rcpsp_sol.get_end_time(task=self.sink_task)
-        penalty = evaluate_constraints(
-            solution=rcpsp_sol, constraints=self.special_constraints
-        )
-        return makespan, penalty
-
-    def evaluate(self, rcpsp_sol: RCPSPSolution) -> Dict[str, float]:
-        obj_makespan, penalty = self.evaluate_function(rcpsp_sol)
-        return {"makespan": obj_makespan, "constraint_penalty": penalty}
-
-    def get_objective_register(self) -> ObjectiveRegister:
-        dict_objective = {
-            "makespan": ObjectiveDoc(type=TypeObjective.OBJECTIVE, default_weight=-1.0),
-            "constraint_penalty": ObjectiveDoc(
-                type=TypeObjective.PENALTY, default_weight=-100.0
-            ),
-        }
-        return ObjectiveRegister(
-            objective_sense=ModeOptim.MAXIMIZATION,
-            objective_handling=ObjectiveHandling.AGGREGATE,
-            dict_objective_to_doc=dict_objective,
-        )
-
-    def satisfy(self, rcpsp_sol: RCPSPSolution) -> bool:
-        s = check_solution(
-            problem=self,
-            solution=rcpsp_sol,
-            relax_the_start_at_end=self.relax_the_start_at_end,
-        )
-        if not s:
-            return s
-        return super().satisfy(rcpsp_sol)
-
-    def get_dummy_solution(self):
-        sol = RCPSPSolutionSpecial(
-            problem=self,
-            rcpsp_permutation=list(range(self.n_jobs_non_dummy)),
-            rcpsp_modes=[1 for i in range(self.n_jobs_non_dummy)],
-        )
-        return sol
-
-    def get_solution_type(self) -> Type[Solution]:
-        return RCPSPSolutionSpecial
 
 
 class RCPSPSolutionSpecialPreemptive(RCPSPSolutionPreemptive):
@@ -583,39 +297,6 @@ class RCPSPModelSpecialConstraintsPreemptive(RCPSPModelPreemptive):
         return RCPSPSolutionSpecialPreemptive
 
 
-def from_rcpsp_model(
-    rcpsp_model: RCPSPModel,
-    constraints: SpecialConstraintsDescription,
-    preemptive=False,
-):
-    if preemptive:
-        return RCPSPModelSpecialConstraintsPreemptive(
-            resources=rcpsp_model.resources,
-            non_renewable_resources=rcpsp_model.non_renewable_resources,
-            mode_details=rcpsp_model.mode_details,
-            successors=rcpsp_model.successors,
-            horizon=rcpsp_model.horizon,
-            special_constraints=constraints,
-            tasks_list=rcpsp_model.tasks_list,
-            source_task=rcpsp_model.source_task,
-            sink_task=rcpsp_model.sink_task,
-            name_task=rcpsp_model.name_task,
-        )
-
-    return RCPSPModelSpecialConstraints(
-        resources=rcpsp_model.resources,
-        non_renewable_resources=rcpsp_model.non_renewable_resources,
-        mode_details=rcpsp_model.mode_details,
-        successors=rcpsp_model.successors,
-        horizon=rcpsp_model.horizon,
-        special_constraints=constraints,
-        tasks_list=rcpsp_model.tasks_list,
-        source_task=rcpsp_model.source_task,
-        sink_task=rcpsp_model.sink_task,
-        name_task=rcpsp_model.name_task,
-    )
-
-
 def evaluate_constraints(
     solution: Union[RCPSPSolution, RCPSPSolutionPreemptive],
     constraints: SpecialConstraintsDescription,
@@ -739,11 +420,8 @@ def compute_constraints_details(
 
 
 def check_solution(
-    problem: Union[
-        RCPSPModelSpecialConstraints, RCPSPModelSpecialConstraintsPreemptive
-    ],
+    problem: RCPSPModelSpecialConstraintsPreemptive,
     solution: Union[
-        RCPSPSolutionSpecial,
         RCPSPSolutionSpecialPreemptive,
         RCPSPSolution,
         RCPSPSolutionPreemptive,
@@ -1575,483 +1253,6 @@ def generate_schedule_from_permutation_serial_sgs_partial_schedule_preempptive(
             rcpsp_schedule[last_act_id] = {}
             rcpsp_schedule[last_act_id]["starts"] = [99999999]
             rcpsp_schedule[last_act_id]["ends"] = [9999999]
-    else:
-        rcpsp_schedule_feasible = True
-    return rcpsp_schedule, rcpsp_schedule_feasible
-
-
-def generate_schedule_from_permutation_serial_sgs(
-    solution: RCPSPSolution, rcpsp_problem: RCPSPModelSpecialConstraints
-):
-    activity_end_times = {}
-
-    unfeasible_non_renewable_resources = False
-    new_horizon = rcpsp_problem.horizon
-    resource_avail_in_time = {}
-    for res in rcpsp_problem.resources_list:
-        if rcpsp_problem.is_varying_resource():
-            resource_avail_in_time[res] = rcpsp_problem.resources[res][
-                : new_horizon + 1
-            ]
-        else:
-            resource_avail_in_time[res] = np.full(
-                new_horizon, rcpsp_problem.resources[res], dtype=np.int_
-            ).tolist()
-    minimum_starting_time = {}
-    for act in rcpsp_problem.tasks_list:
-        minimum_starting_time[act] = 0
-        if rcpsp_problem.do_special_constraints:
-            if act in rcpsp_problem.special_constraints.start_times_window:
-                minimum_starting_time[act] = (
-                    rcpsp_problem.special_constraints.start_times_window[act][0]
-                    if rcpsp_problem.special_constraints.start_times_window[act][0]
-                    is not None
-                    else 0
-                )
-    perm_extended = [
-        rcpsp_problem.tasks_list_non_dummy[x] for x in solution.rcpsp_permutation
-    ]
-    perm_extended.insert(0, rcpsp_problem.source_task)
-    perm_extended.append(rcpsp_problem.sink_task)
-    modes_dict = rcpsp_problem.build_mode_dict(solution.rcpsp_modes)
-
-    def ressource_consumption(res, task, duration, mode):
-        dur = rcpsp_problem.mode_details[task][mode]["duration"]
-        if duration > dur:
-            return 0
-        return rcpsp_problem.mode_details[task][mode].get(res, 0)
-
-    for k in modes_dict:
-        if modes_dict[k] not in rcpsp_problem.mode_details[k]:
-            modes_dict[k] = 1
-
-    def look_for_task(perm, ignore_sc=False):
-        act_ids = []
-        for task_id in perm:
-            respected = True
-            # Check all kind of precedence constraints....
-            for pred in rcpsp_problem.predecessors.get(task_id, {}):
-                if pred in perm_extended:
-                    respected = False
-                    break
-            if not ignore_sc:
-                for (
-                    pred
-                ) in rcpsp_problem.special_constraints.dict_start_at_end_reverse.get(
-                    task_id, {}
-                ):
-                    if pred in perm_extended:
-                        respected = False
-                        break
-                for (
-                    pred
-                ) in rcpsp_problem.special_constraints.dict_start_at_end_offset_reverse.get(
-                    task_id, {}
-                ):
-                    if pred in perm_extended:
-                        respected = False
-                        break
-                for (
-                    pred
-                ) in rcpsp_problem.special_constraints.dict_start_after_nunit_reverse.get(
-                    task_id, {}
-                ):
-                    if pred in perm_extended:
-                        respected = False
-                        break
-            task_to_start_too = set()
-            if respected:
-                task_to_start_too = (
-                    rcpsp_problem.special_constraints.dict_start_together.get(
-                        task_id, set()
-                    )
-                )
-                if not ignore_sc:
-                    if len(task_to_start_too) > 0:
-                        if not all(
-                            s not in perm_extended
-                            for t in task_to_start_too
-                            for s in rcpsp_problem.predecessors[t]
-                        ):
-                            respected = False
-            if respected:
-                act_ids = [task_id] + list(task_to_start_too)
-                break
-        return act_ids
-
-    while len(perm_extended) > 0 and not unfeasible_non_renewable_resources:
-        act_ids = look_for_task(
-            [
-                k
-                for k in rcpsp_problem.special_constraints.dict_start_at_end_reverse
-                if k in perm_extended
-            ]
-        )
-        act_ids = []
-        if len(act_ids) == 0:
-            act_ids = look_for_task(perm_extended)
-        if len(act_ids) == 0:
-            act_ids = look_for_task(perm_extended, ignore_sc=True)
-        current_min_time = max([minimum_starting_time[act_id] for act_id in act_ids])
-        max_duration = max(
-            [
-                rcpsp_problem.mode_details[act_id][modes_dict[act_id]]["duration"]
-                for act_id in act_ids
-            ]
-        )
-        valid = False
-        while not valid:
-            valid = True
-            for t in range(current_min_time, current_min_time + max_duration):
-                for res in rcpsp_problem.resources_list:
-                    r = sum(
-                        [
-                            ressource_consumption(
-                                res=res,
-                                task=task,
-                                duration=t - current_min_time,
-                                mode=modes_dict[task],
-                            )
-                            for task in act_ids
-                        ]
-                    )
-                    if r == 0:
-                        continue
-                    if t < new_horizon:
-                        if resource_avail_in_time[res][t] < r:
-                            valid = False
-                            break
-                    else:
-                        unfeasible_non_renewable_resources = True
-                if not valid:
-                    break
-            if not valid:
-                current_min_time += 1
-        if not unfeasible_non_renewable_resources:
-            end_t = current_min_time + max_duration
-            for t in range(current_min_time, current_min_time + max_duration):
-                for res in resource_avail_in_time:
-                    r = sum(
-                        [
-                            ressource_consumption(
-                                res=res,
-                                task=task,
-                                duration=t - current_min_time,
-                                mode=modes_dict[task],
-                            )
-                            for task in act_ids
-                        ]
-                    )
-                    resource_avail_in_time[res][t] -= r
-                    if res in rcpsp_problem.non_renewable_resources and t == end_t - 1:
-                        for tt in range(end_t, new_horizon):
-                            resource_avail_in_time[res][tt] -= r
-                            if resource_avail_in_time[res][tt] < 0:
-                                unfeasible_non_renewable_resources = True
-            for act_id in act_ids:
-                activity_end_times[act_id] = (
-                    current_min_time
-                    + rcpsp_problem.mode_details[act_id][modes_dict[act_id]]["duration"]
-                )
-                perm_extended.remove(act_id)
-                for s in rcpsp_problem.successors[act_id]:
-                    minimum_starting_time[s] = max(
-                        minimum_starting_time[s], activity_end_times[act_id]
-                    )
-                for s in rcpsp_problem.special_constraints.dict_start_at_end.get(
-                    act_id, {}
-                ):
-                    minimum_starting_time[s] = max(
-                        minimum_starting_time[s], activity_end_times[act_id]
-                    )
-                for s in rcpsp_problem.special_constraints.dict_start_after_nunit.get(
-                    act_id, {}
-                ):
-                    minimum_starting_time[s] = max(
-                        minimum_starting_time[s],
-                        current_min_time
-                        + rcpsp_problem.special_constraints.dict_start_after_nunit[
-                            act_id
-                        ][s],
-                    )
-                for s in rcpsp_problem.special_constraints.dict_start_at_end_offset.get(
-                    act_id, {}
-                ):
-                    minimum_starting_time[s] = max(
-                        minimum_starting_time[s],
-                        activity_end_times[act_id]
-                        + rcpsp_problem.special_constraints.dict_start_at_end_offset[
-                            act_id
-                        ][s],
-                    )
-    rcpsp_schedule = {}
-    for act_id in activity_end_times:
-        rcpsp_schedule[act_id] = {}
-        rcpsp_schedule[act_id]["start_time"] = (
-            activity_end_times[act_id]
-            - rcpsp_problem.mode_details[act_id][modes_dict[act_id]]["duration"]
-        )
-        rcpsp_schedule[act_id]["end_time"] = activity_end_times[act_id]
-    if unfeasible_non_renewable_resources:
-        rcpsp_schedule_feasible = False
-        last_act_id = rcpsp_problem.sink_task
-        if last_act_id not in rcpsp_schedule:
-            rcpsp_schedule[last_act_id] = {}
-            rcpsp_schedule[last_act_id]["start_time"] = 99999999
-            rcpsp_schedule[last_act_id]["end_time"] = 9999999
-    else:
-        rcpsp_schedule_feasible = True
-    return rcpsp_schedule, rcpsp_schedule_feasible
-
-
-def generate_schedule_from_permutation_serial_sgs_partial_schedule(
-    solution: RCPSPSolution,
-    rcpsp_problem: RCPSPModelSpecialConstraints,
-    current_t,
-    completed_tasks,
-    scheduled_tasks_start_times,
-):
-    activity_end_times = {}
-    unfeasible_non_renewable_resources = False
-    new_horizon = rcpsp_problem.horizon
-    resource_avail_in_time = {}
-    for res in rcpsp_problem.resources_list:
-        if rcpsp_problem.is_varying_resource():
-            resource_avail_in_time[res] = rcpsp_problem.resources[res][
-                : new_horizon + 1
-            ]
-        else:
-            resource_avail_in_time[res] = np.full(
-                new_horizon, rcpsp_problem.resources[res], dtype=np.int_
-            ).tolist()
-
-    def ressource_consumption(res, task, duration, mode):
-        dur = rcpsp_problem.mode_details[task][mode]["duration"]
-        if duration > dur:
-            return 0
-        return rcpsp_problem.mode_details[task][mode].get(res, 0)
-
-    minimum_starting_time = {}
-    for act in rcpsp_problem.tasks_list:
-        if act in list(scheduled_tasks_start_times.keys()):
-            minimum_starting_time[act] = scheduled_tasks_start_times[act]
-        else:
-            minimum_starting_time[act] = current_t
-        if rcpsp_problem.do_special_constraints:
-            if act in rcpsp_problem.special_constraints.start_times_window:
-                minimum_starting_time[act] = (
-                    max(
-                        rcpsp_problem.special_constraints.start_times_window[act][0],
-                        minimum_starting_time[act],
-                    )
-                    if rcpsp_problem.special_constraints.start_times_window[act][0]
-                    is not None
-                    else minimum_starting_time[act]
-                )
-    perm_extended = [
-        rcpsp_problem.tasks_list_non_dummy[x] for x in solution.rcpsp_permutation
-    ]
-    perm_extended.insert(0, rcpsp_problem.source_task)
-    perm_extended.append(rcpsp_problem.sink_task)
-    modes_dict = rcpsp_problem.build_mode_dict(solution.rcpsp_modes)
-    # Update current resource usage by the scheduled task (ongoing task, in practice)
-    for act_id in scheduled_tasks_start_times:
-        current_min_time = scheduled_tasks_start_times[act_id]
-        end_t = (
-            current_min_time
-            + rcpsp_problem.mode_details[act_id][modes_dict[act_id]]["duration"]
-        )
-        for t in range(current_min_time, end_t):
-            for res in resource_avail_in_time:
-                resource_avail_in_time[res][t] -= rcpsp_problem.mode_details[act_id][
-                    modes_dict[act_id]
-                ].get(res, 0)
-                if res in rcpsp_problem.non_renewable_resources and t == end_t - 1:
-                    for tt in range(end_t, new_horizon):
-                        resource_avail_in_time[res][tt] -= rcpsp_problem.mode_details[
-                            act_id
-                        ][modes_dict[act_id]].get(res, 0)
-                        if resource_avail_in_time[res][tt] < 0:
-                            unfeasible_non_renewable_resources = True
-        activity_end_times[act_id] = end_t
-        perm_extended.remove(act_id)
-        for s in rcpsp_problem.successors[act_id]:
-            minimum_starting_time[s] = max(
-                minimum_starting_time[s], activity_end_times[act_id]
-            )
-
-    perm_extended = [x for x in perm_extended if x not in list(completed_tasks)]
-    for ac in modes_dict:
-        if modes_dict[ac] not in rcpsp_problem.mode_details[ac]:
-            modes_dict[ac] = 1
-
-    while len(perm_extended) > 0 and not unfeasible_non_renewable_resources:
-        act_ids = []
-        for task_id in perm_extended:
-            respected = True
-            # Check all kind of precedence constraints....
-            for pred in rcpsp_problem.predecessors.get(task_id, {}):
-                if pred in perm_extended:
-                    respected = False
-                    break
-            for pred in rcpsp_problem.special_constraints.dict_start_at_end_reverse.get(
-                task_id, {}
-            ):
-                if pred in perm_extended:
-                    respected = False
-                    break
-            for (
-                pred
-            ) in rcpsp_problem.special_constraints.dict_start_at_end_offset_reverse.get(
-                task_id, {}
-            ):
-                if pred in perm_extended:
-                    respected = False
-                    break
-            for (
-                pred
-            ) in rcpsp_problem.special_constraints.dict_start_after_nunit_reverse.get(
-                task_id, {}
-            ):
-                if pred in perm_extended:
-                    respected = False
-                    break
-            task_to_start_too = set()
-            if respected:
-                task_to_start_too = [
-                    k
-                    for k in rcpsp_problem.special_constraints.dict_start_together.get(
-                        task_id, set()
-                    )
-                    if k in perm_extended
-                ]
-                if len(task_to_start_too) > 0:
-                    if not all(
-                        s not in perm_extended
-                        for t in task_to_start_too
-                        for s in rcpsp_problem.predecessors[t]
-                    ):
-                        respected = False
-            if respected:
-                act_ids = [task_id] + list(task_to_start_too)
-                break
-        if len(act_ids) == 0:
-            for task_id in perm_extended:
-                respected = True
-                # Check all kind of precedence constraints....
-                for pred in rcpsp_problem.predecessors.get(task_id, {}):
-                    if pred in perm_extended:
-                        respected = False
-                        break
-                if respected:
-                    act_ids = [task_id]
-        current_min_time = max([minimum_starting_time[act_id] for act_id in act_ids])
-        max_duration = max(
-            [
-                rcpsp_problem.mode_details[act_id][modes_dict[act_id]]["duration"]
-                for act_id in act_ids
-            ]
-        )
-        valid = False
-        while not valid:
-            valid = True
-            for t in range(current_min_time, current_min_time + max_duration):
-                for res in rcpsp_problem.resources_list:
-                    r = sum(
-                        [
-                            ressource_consumption(
-                                res=res,
-                                task=task,
-                                duration=t - current_min_time,
-                                mode=modes_dict[task],
-                            )
-                            for task in act_ids
-                        ]
-                    )
-                    if r == 0:
-                        continue
-                    if t < new_horizon:
-                        if resource_avail_in_time[res][t] < r:
-                            valid = False
-                            break
-                    else:
-                        unfeasible_non_renewable_resources = True
-                if not valid:
-                    break
-            if not valid:
-                current_min_time += 1
-        if not unfeasible_non_renewable_resources:
-            end_t = current_min_time + max_duration
-            for t in range(current_min_time, current_min_time + max_duration):
-                for res in resource_avail_in_time:
-                    r = sum(
-                        [
-                            ressource_consumption(
-                                res=res,
-                                task=task,
-                                duration=t - current_min_time,
-                                mode=modes_dict[task],
-                            )
-                            for task in act_ids
-                        ]
-                    )
-                    resource_avail_in_time[res][t] -= r
-                    if res in rcpsp_problem.non_renewable_resources and t == end_t - 1:
-                        for tt in range(end_t, new_horizon):
-                            resource_avail_in_time[res][tt] -= r
-                            if resource_avail_in_time[res][tt] < 0:
-                                unfeasible_non_renewable_resources = True
-            for act_id in act_ids:
-                activity_end_times[act_id] = (
-                    current_min_time
-                    + rcpsp_problem.mode_details[act_id][modes_dict[act_id]]["duration"]
-                )
-                perm_extended.remove(act_id)
-                for s in rcpsp_problem.successors[act_id]:
-                    minimum_starting_time[s] = max(
-                        minimum_starting_time[s], activity_end_times[act_id]
-                    )
-                for s in rcpsp_problem.special_constraints.dict_start_at_end.get(
-                    act_id, {}
-                ):
-                    minimum_starting_time[s] = activity_end_times[act_id]
-                for s in rcpsp_problem.special_constraints.dict_start_after_nunit.get(
-                    act_id, {}
-                ):
-                    minimum_starting_time[s] = (
-                        current_min_time
-                        + rcpsp_problem.special_constraints.dict_start_after_nunit[
-                            act_id
-                        ][s]
-                    )
-                for s in rcpsp_problem.special_constraints.dict_start_at_end_offset.get(
-                    act_id, {}
-                ):
-                    minimum_starting_time[s] = (
-                        activity_end_times[act_id]
-                        + rcpsp_problem.special_constraints.dict_start_at_end_offset[
-                            act_id
-                        ][s]
-                    )
-    rcpsp_schedule = {}
-    for act_id in activity_end_times:
-        rcpsp_schedule[act_id] = {}
-        rcpsp_schedule[act_id]["start_time"] = (
-            activity_end_times[act_id]
-            - rcpsp_problem.mode_details[act_id][modes_dict[act_id]]["duration"]
-        )
-        rcpsp_schedule[act_id]["end_time"] = activity_end_times[act_id]
-    for act_id in completed_tasks:
-        rcpsp_schedule[act_id] = {}
-        rcpsp_schedule[act_id]["start_time"] = completed_tasks[act_id].start
-        rcpsp_schedule[act_id]["end_time"] = completed_tasks[act_id].end
-    if unfeasible_non_renewable_resources:
-        rcpsp_schedule_feasible = False
-        last_act_id = rcpsp_problem.sink_task
-        if last_act_id not in rcpsp_schedule:
-            rcpsp_schedule[last_act_id] = {}
-            rcpsp_schedule[last_act_id]["start_time"] = 99999999
-            rcpsp_schedule[last_act_id]["end_time"] = 9999999
     else:
         rcpsp_schedule_feasible = True
     return rcpsp_schedule, rcpsp_schedule_feasible
