+++ discrete_optimization/rcpsp_multiskill/solvers/cp_solvers.py
@@ -29,11 +29,8 @@ from discrete_optimization.generic_tools.do_problem import (
 from discrete_optimization.generic_tools.result_storage.result_storage import (
     ResultStorage,
 )
-from discrete_optimization.rcpsp.rcpsp_model import (
-    PartialSolution,
-    RCPSPModel,
-    RCPSPSolution,
-)
+from discrete_optimization.rcpsp.core import PartialSolution
+from discrete_optimization.rcpsp.rcpsp_model import RCPSPModel, RCPSPSolution
 from discrete_optimization.rcpsp.rcpsp_model_preemptive import RCPSPSolutionPreemptive
 from discrete_optimization.rcpsp_multiskill.rcpsp_multiskill import (
     MS_RCPSPModel,
